+++ my-app/src/tests/App.test.jsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from '../views/App';
 
 test('renders learn react link', () => {
   render(<App />);
